# Upgrade expected attention with support for more models